chatbot-be/pyproject.toml
[project]
name = "chatbot-be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.123.0",
    "uvicorn[standard]==0.38.0",
    "pydantic==2.12.4",
    "pydantic-settings==2.12.0",
    "requests==2.32.5",
    "sqlalchemy==2.0.44",
    "alembic==1.17.2",
    "python-dotenv==1.2.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "apscheduler==3.11.1",
    "pypdf==6.1.3",
    "langchain==1.0.1",
    "langchain-classic==1.0.0",
    "langchain-pinecone==0.2.13",
    "langchain-openai==1.0.1",
    "langchain-community==0.4",
    "langchain-huggingface==1.2.0",
    "sentence-transformers==5.1.2",
]
